requirements.txt
python-dotenv
requests
slack-sdk
